[ISSUE #9872]♻️Split ControllerManager into focused modules
